add callback function in the acmp machine and change connect table connect and diconnect function interface

+++ lib/include/message_func/func_proccess.h
@@ -61,7 +61,7 @@ typedef struct proccess_func_items
 
 extern fcwqueue fcwork_queue;		// 函数命令消息工作队列
 
-extern uint16_t get_sys_state( void );
+extern uint8_t get_sys_state( void );
 extern void init_func_command_work_queue( void );
 extern int func_command_find_and_run( proccess_func_items * func_tables, fcwqueue*  p_func_wq );
 extern bool find_func_command_link( uint8_t user, uint16_t cfc_cmd, uint16_t func_cmd, const uint8_t *pdata, const uint16_t data_len );
